+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 dev:
   - provide support for additional S3 store options
+  - show error when attempting to delete non-filesystem wallets
 
 1.26.2
   - remove check that requires capella prior to generating validator credentials change operations